refactor(web): extract UI components from SettingsPage
Extracted the following components from web/src/pages/SettingsPage.tsx:
- DynamicProviderCard -> web/src/components/DynamicProviderCard.tsx
- GeneralTab -> web/src/components/GeneralSettingsTab.tsx
- CircuitBreakersTab -> web/src/components/CircuitBreakersTab.tsx

All functionality and styles remain identical. Prop interfaces are strongly typed.
