+++ core/npm_core_dev_output.log
@@ -1,0 +1,12 @@
+
+> core@1.0.0 dev
+> bunx tsx watch src/index.ts
+
+sh: 1: bunx: not found
+npm error Lifecycle script `dev` failed with error:
+npm error code 127
+npm error path /app/core
+npm error workspace core@1.0.0
+npm error location /app/core
+npm error command failed
+npm error command sh -c bunx tsx watch src/index.ts
