+++ core/npm_core_start_output.log
@@ -1,0 +1,119 @@
+
+> core@1.0.0 start
+> node dist/index.js
+
+[IdentityService] JWT_SECRET not set — using a random ephemeral secret. Tokens will not survive server restarts. Set JWT_SECRET in production.
+[AuthService] Registered auth plugin: api-key
+[SERACore] OIDC_ISSUER_URL not set — running in API-key-only mode
+[ProviderRegistry] No providers config at /app/config/providers.json — relying on env-var bootstrap and auto-detection
+[DynamicProviderManager] No dynamic providers config at /app/config/dynamic_providers.json
+[AgentManifestLoader] loadAllManifests: 0.077ms
+[Orchestrator] Loaded 0 agent templates
+[AgentManifestLoader] loadAllManifests: 0.037ms
+[MCPRegistry] Registered embedded sera-core MCP tools
+[MCPServerManager] Spawning MCP server container: filesystem (mcp-filesystem-8234eac1)
+[SERACore] Failed to load MCP servers: Error: EACCES: permission denied, mkdir '/memory/mcp-filesystem-8234eac1'
+    at Object.mkdirSync (node:fs:1370:26)
+    at SandboxManager.spawn (file:///app/core/dist/sandbox/SandboxManager.js:110:8)
+    at MCPServerManager.spawnServer (file:///app/core/dist/mcp/MCPServerManager.js:32:44)
+    at MCPRegistry.registerContainerServer (file:///app/core/dist/mcp/registry.js:59:56)
+    at MCPRegistry.loadManifest (file:///app/core/dist/mcp/registry.js:139:16)
+    at MCPRegistry.loadFromDirectory (file:///app/core/dist/mcp/registry.js:99:20)
+    at startServer (file:///app/core/dist/index.js:398:21) {
+  errno: -13,
+  code: 'EACCES',
+  syscall: 'mkdir',
+  path: '/memory/mcp-filesystem-8234eac1'
+}
+[Orchestrator] Watching for agent changes in /app/agents
+(node:9276) [DEP0040] DeprecationWarning: The `punycode` module is deprecated. Please use a userland alternative instead.
+(Use `node --trace-deprecation ...` to show where the warning was created)
+[Intercom] Failed to publish to system.mcp_registry_update: getaddrinfo ENOTFOUND centrifugo
+Failed to obtain server version. Unable to check client-server compatibility. Set checkCompatibility=false to skip version check.
+Failed to obtain server version. Unable to check client-server compatibility. Set checkCompatibility=false to skip version check.
+Failed to obtain server version. Unable to check client-server compatibility. Set checkCompatibility=false to skip version check.
+Failed to obtain server version. Unable to check client-server compatibility. Set checkCompatibility=false to skip version check.
+Failed to obtain server version. Unable to check client-server compatibility. Set checkCompatibility=false to skip version check.
+Failed to obtain server version. Unable to check client-server compatibility. Set checkCompatibility=false to skip version check.
+could not connect to postgres: AggregateError [ECONNREFUSED]:
+    at internalConnectMultiple (node:net:1134:18)
+    at afterConnectMultiple (node:net:1715:7) {
+  code: 'ECONNREFUSED',
+  [errors]: [
+    Error: connect ECONNREFUSED ::1:5432
+        at createConnectionError (node:net:1678:14)
+        at afterConnectMultiple (node:net:1708:16) {
+      errno: -111,
+      code: 'ECONNREFUSED',
+      syscall: 'connect',
+      address: '::1',
+      port: 5432
+    },
+    Error: connect ECONNREFUSED 127.0.0.1:5432
+        at createConnectionError (node:net:1678:14)
+        at afterConnectMultiple (node:net:1708:16) {
+      errno: -111,
+      code: 'ECONNREFUSED',
+      syscall: 'connect',
+      address: '127.0.0.1',
+      port: 5432
+    }
+  ]
+}
+[Database] Database migration failed: AggregateError [ECONNREFUSED]:
+    at internalConnectMultiple (node:net:1134:18)
+    at afterConnectMultiple (node:net:1715:7) {
+  code: 'ECONNREFUSED',
+  [errors]: [
+    Error: connect ECONNREFUSED ::1:5432
+        at createConnectionError (node:net:1678:14)
+        at afterConnectMultiple (node:net:1708:16) {
+      errno: -111,
+      code: 'ECONNREFUSED',
+      syscall: 'connect',
+      address: '::1',
+      port: 5432
+    },
+    Error: connect ECONNREFUSED 127.0.0.1:5432
+        at createConnectionError (node:net:1678:14)
+        at afterConnectMultiple (node:net:1708:16) {
+      errno: -111,
+      code: 'ECONNREFUSED',
+      syscall: 'connect',
+      address: '127.0.0.1',
+      port: 5432
+    }
+  ]
+}
+[SERACore] Failed to start server: AggregateError [ECONNREFUSED]:
+    at internalConnectMultiple (node:net:1134:18)
+    at afterConnectMultiple (node:net:1715:7) {
+  code: 'ECONNREFUSED',
+  [errors]: [
+    Error: connect ECONNREFUSED ::1:5432
+        at createConnectionError (node:net:1678:14)
+        at afterConnectMultiple (node:net:1708:16) {
+      errno: -111,
+      code: 'ECONNREFUSED',
+      syscall: 'connect',
+      address: '::1',
+      port: 5432
+    },
+    Error: connect ECONNREFUSED 127.0.0.1:5432
+        at createConnectionError (node:net:1678:14)
+        at afterConnectMultiple (node:net:1708:16) {
+      errno: -111,
+      code: 'ECONNREFUSED',
+      syscall: 'connect',
+      address: '127.0.0.1',
+      port: 5432
+    }
+  ]
+}
+npm error Lifecycle script `start` failed with error:
+npm error code 1
+npm error path /app/core
+npm error workspace core@1.0.0
+npm error location /app/core
+npm error command failed
+npm error command sh -c node dist/index.js
