+++ web/src/pages/SettingsPage.tsx
@@ -1,19 +1,15 @@
-import { useState, useEffect } from 'react';
+import { useState } from 'react';
 import {
   Zap,
   CheckCircle,
   XCircle,
   RefreshCw,
-  ChevronDown,
-  ChevronUp,
   Radio,
   Layers,
   Settings2,
   Sliders,
   Activity,
   Plus,
-  Trash2,
-  ExternalLink,
   Save,
 } from 'lucide-react';
 import {
@@ -27,133 +23,12 @@ import {
 import { useCircuitBreakers, useResetCircuitBreaker } from '@/hooks/useHealth';
 import * as providersApi from '@/lib/api/providers';
 import { Spinner } from '@/components/ui/spinner';
-import { Badge } from '@/components/ui/badge';
 import { Button } from '@/components/ui/button';
+import { DynamicProviderCard } from '@/components/DynamicProviderCard';
+import { GeneralTab } from '@/components/GeneralSettingsTab';
+import { CircuitBreakersTab } from '@/components/CircuitBreakersTab';
 
 type Tab = 'providers' | 'models' | 'general' | 'circuit-breakers';
-
-function DynamicProviderCard({
-  provider,
-  status,
-  onRemove,
-}: {
-  provider: providersApi.DynamicProviderConfig;
-  status?: providersApi.DynamicProviderStatus;
-  onRemove: (id: string) => void;
-}) {
-  const [expanded, setExpanded] = useState(false);
-  const isHealthy = status?.status === 'ok';
-
-  return (
-    <div
-      className={`sera-card-static overflow-hidden ${
-        isHealthy
-          ? 'border-sera-success/30'
-          : 'border-sera-error/30 shadow-[0_0_15px_rgba(255,82,82,0.05)]'
-      }`}
-    >
-      <button
-        onClick={() => setExpanded((e) => !e)}
-        className="w-full p-4 flex items-center justify-between hover:bg-sera-surface-hover transition-colors"
-      >
-        <div className="flex items-center gap-3">
-          <div className="w-9 h-9 rounded-lg flex items-center justify-center bg-amber-500/10 border border-amber-500/20">
-            <Radio size={16} className="text-amber-400" />
-          </div>
-          <div className="text-left">
-            <div className="flex items-center gap-2">
-              <h3 className="text-sm font-semibold text-sera-text">{provider.name}</h3>
-              {isHealthy ? (
-                <Badge variant="success" className="text-[9px] px-1.5 py-0">
-                  Online
-                </Badge>
-              ) : (
-                <Badge variant="error" className="text-[9px] px-1.5 py-0">
-                  Offline
-                </Badge>
-              )}
-            </div>
-            <p className="text-[11px] text-sera-text-muted mt-0.5 font-mono select-all">
-              {provider.baseUrl}
-            </p>
-          </div>
-        </div>
-        <div className="flex items-center gap-3">
-          <span className="text-[11px] text-sera-text-dim px-2 py-0.5 rounded-full bg-sera-bg/80 border border-sera-border">
-            {status?.discoveredModels?.length ?? 0} models
-          </span>
-          {expanded ? (
-            <ChevronUp size={14} className="text-sera-text-dim" />
-          ) : (
-            <ChevronDown size={14} className="text-sera-text-dim" />
-          )}
-        </div>
-      </button>
-
-      {expanded && (
-        <div className="border-t border-sera-border p-4 space-y-4 bg-sera-bg/50 animate-in slide-in-from-top-2 duration-200">
-          <div className="space-y-3">
-            <div className="flex justify-between items-center px-1">
-              <span className="text-[11px] font-medium text-sera-text-dim uppercase tracking-wider flex items-center gap-1.5">
-                <Activity size={10} /> Discovery Status
-              </span>
-              <span className="text-[11px] text-sera-text-muted font-mono bg-sera-surface/80 px-2 py-0.5 rounded border border-sera-border">
-                {status?.lastCheck ? new Date(status.lastCheck).toLocaleTimeString() : 'Never'}
-              </span>
-            </div>
-
-            {!isHealthy && status?.error && (
-              <div className="flex items-start gap-2 p-3 rounded-lg bg-sera-error/5 border border-sera-error/20 text-sera-error text-[11px] leading-relaxed">
-                <XCircle size={14} className="mt-0.5 shrink-0" />
-                <span>{status.error}</span>
-              </div>
-            )}
-
-            {isHealthy && (
-              <div className="space-y-2">
-                <span className="text-[11px] text-sera-text-dim px-1 block">Live Models:</span>
-                <div className="flex flex-wrap gap-2 p-2 rounded-lg bg-sera-bg/30 border border-sera-border/50">
-                  {status?.discoveredModels?.map((m) => (
-                    <span
-                      key={m}
-                      className="px-2 py-1 rounded border border-sera-border bg-sera-surface/50 text-[10px] text-sera-text-muted font-mono hover:border-sera-accent/30 hover:text-sera-text transition-colors cursor-default"
-                    >
-                      {m}
-                    </span>
-                  ))}
-                  {status?.discoveredModels?.length === 0 && (
-                    <span className="text-[11px] text-sera-text-dim italic px-2 py-1">
-                      No models found — start them in LM Studio
-                    </span>
-                  )}
-                </div>
-              </div>
-            )}
-          </div>
-
-          <div className="flex gap-2 border-t border-sera-border pt-4 mt-2">
-            <Button
-              variant="outline"
-              size="sm"
-              className="flex-1 text-xs h-9 bg-sera-error/5 hover:bg-sera-error/10 text-sera-error border-sera-error/20 gap-2"
-              onClick={() => onRemove(provider.id)}
-            >
-              <Trash2 size={13} /> Remove
-            </Button>
-            <Button
-              variant="outline"
-              size="sm"
-              className="flex-1 text-xs h-9 gap-2"
-              onClick={() => window.open(provider.baseUrl, '_blank')}
-            >
-              <ExternalLink size={13} /> API Info
-            </Button>
-          </div>
-        </div>
-      )}
-    </div>
-  );
-}
 
 export default function SettingsPage() {
   const [tab, setTab] = useState<Tab>('providers');
@@ -523,172 +398,3 @@ export default function SettingsPage() {
   );
 }
 
-function GeneralTab({
-  registeredModels,
-}: {
-  registeredModels: Array<{ modelName: string; description?: string; provider?: string }>;
-}) {
-  const [defaultModel, setDefaultModelLocal] = useState('');
-  const [saving, setSaving] = useState(false);
-  const [loaded, setLoaded] = useState(false);
-
-  useEffect(() => {
-    providersApi
-      .getDefaultModel()
-      .then((res) => {
-        if (res.defaultModel) setDefaultModelLocal(res.defaultModel);
-        setLoaded(true);
-      })
-      .catch(() => setLoaded(true));
-  }, []);
-
-  return (
-    <div className="sera-card-static p-6 space-y-6 max-w-xl animate-in fade-in slide-in-from-bottom-2 duration-300">
-      <div>
-        <h3 className="text-xs font-semibold tracking-[0.1em] text-sera-text-dim uppercase mb-4">
-          Default Model
-        </h3>
-        <p className="text-xs text-sera-text-muted mb-3">
-          Agents configured with &ldquo;default&rdquo; as their model will use this model.
-        </p>
-        <div className="flex gap-3 items-end">
-          <div className="flex-1 space-y-1.5">
-            <label className="text-[11px] font-medium text-sera-text-dim uppercase tracking-wider">
-              Model
-            </label>
-            {loaded ? (
-              <select
-                value={defaultModel}
-                onChange={(e) => setDefaultModelLocal(e.target.value)}
-                className="sera-input text-xs"
-              >
-                <option value="">— No default —</option>
-                {registeredModels.map((m) => (
-                  <option key={m.modelName} value={m.modelName}>
-                    {m.description ?? m.modelName}
-                  </option>
-                ))}
-              </select>
-            ) : (
-              <div className="h-9 bg-sera-surface rounded animate-pulse" />
-            )}
-          </div>
-          <Button
-            size="sm"
-            className="h-9 text-xs gap-1.5 bg-sera-accent hover:bg-sera-accent-hover text-sera-bg"
-            disabled={saving || !defaultModel}
-            onClick={() => {
-              setSaving(true);
-              providersApi
-                .setDefaultModel(defaultModel)
-                .then(() => setSaving(false))
-                .catch(() => setSaving(false));
-            }}
-          >
-            <Save size={13} /> Save
-          </Button>
-        </div>
-      </div>
-      <div className="border-t border-sera-border pt-6">
-        <h3 className="text-xs font-semibold tracking-[0.1em] text-sera-text-dim uppercase mb-3">
-          System Info
-        </h3>
-        <div className="space-y-2 text-xs">
-          <div className="flex justify-between">
-            <span className="text-sera-text-muted">Platform</span>
-            <span className="text-sera-text">SERA v1.0</span>
-          </div>
-          <div className="flex justify-between">
-            <span className="text-sera-text-muted">Runtime</span>
-            <span className="text-sera-text">Node.js 22 + TypeScript</span>
-          </div>
-          <div className="flex justify-between">
-            <span className="text-sera-text-muted">Frontend</span>
-            <span className="text-sera-text">Vite + React Router v7</span>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function cbStateBadge(state: string) {
-  if (state === 'open') return <Badge variant="error">Open</Badge>;
-  if (state === 'half-open') return <Badge variant="warning">Half-Open</Badge>;
-  return <Badge variant="success">Closed</Badge>;
-}
-
-function CircuitBreakersTab({
-  breakers,
-  onReset,
-  resetting,
-}: {
-  breakers: Array<{
-    provider: string;
-    state: string;
-    failures: number;
-    lastFailureAt?: string;
-    nextRetryAt?: string;
-  }>;
-  onReset: (provider: string) => void;
-  resetting: boolean;
-}) {
-  return (
-    <div className="space-y-4">
-      <p className="text-sm text-sera-text-muted">
-        Circuit breakers protect against repeated failures. When open, requests to the provider are
-        paused.
-      </p>
-      {breakers.length === 0 ? (
-        <div className="sera-card-static p-8 text-center text-sera-text-dim text-sm">
-          No circuit breaker data — all providers healthy.
-        </div>
-      ) : (
-        <div className="sera-card-static overflow-hidden">
-          <table className="w-full text-sm">
-            <thead>
-              <tr className="border-b border-sera-border text-[11px] uppercase tracking-wider text-sera-text-dim">
-                <th className="text-left py-3 px-4">Provider</th>
-                <th className="text-left py-3 px-4">State</th>
-                <th className="text-left py-3 px-4">Failures</th>
-                <th className="text-left py-3 px-4">Last Failure</th>
-                <th className="text-left py-3 px-4">Next Retry</th>
-                <th className="py-3 px-4" />
-              </tr>
-            </thead>
-            <tbody>
-              {breakers.map((cb) => (
-                <tr
-                  key={cb.provider}
-                  className="border-b border-sera-border/50 hover:bg-sera-surface-hover transition-colors"
-                >
-                  <td className="py-3 px-4 font-mono text-xs text-sera-text">{cb.provider}</td>
-                  <td className="py-3 px-4">{cbStateBadge(cb.state)}</td>
-                  <td className="py-3 px-4 text-sera-text-muted">{cb.failures}</td>
-                  <td className="py-3 px-4 text-xs text-sera-text-muted">
-                    {cb.lastFailureAt ? new Date(cb.lastFailureAt).toLocaleString() : '—'}
-                  </td>
-                  <td className="py-3 px-4 text-xs text-sera-text-muted">
-                    {cb.nextRetryAt ? new Date(cb.nextRetryAt).toLocaleString() : '—'}
-                  </td>
-                  <td className="py-3 px-4 text-right">
-                    {cb.state !== 'closed' && (
-                      <Button
-                        size="sm"
-                        variant="ghost"
-                        disabled={resetting}
-                        onClick={() => onReset(cb.provider)}
-                      >
-                        <RefreshCw size={12} /> Reset
-                      </Button>
-                    )}
-                  </td>
-                </tr>
-              ))}
-            </tbody>
-          </table>
-        </div>
-      )}
-    </div>
-  );
-}
